pyproject.toml
[tool.poetry]
name = "11labs-bot"
version = "0.1.0"
description = "AI voice bot for Discord"
authors = ["me"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.0"
elevenlabs = "^0.2.26"
pynacl = "^1.5.0"
ffmpeg = "^1.4"
ffmpeg-python = "^0.2.0"
python-ffmpeg = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
